Replace boolean / checkbox with combo box
Although I have no way to test it it seems that FFmpeg supports other
hardware decode methods which can be easily selected by using results
of `av_hwdevice_find_type_by_name`. I don't really have a way to test
it though, as vdpau vaapi engine crashes my laptop (not only in context
of Chiaki, but also when I try to use it with other apps).

Working on tests. And I definitely need to work on naming structures and
variables, I'm really bad at it (that said, I'm going to iterate on this
code).